+++ src/extension.ts
@@ -70,14 +70,22 @@ import { registerLocalResourceWorkflows } from "./docker/workflows/workflowIniti
 import { MessageDocumentProvider } from "./documentProviders/message";
 import { SchemaDocumentProvider } from "./documentProviders/schema";
 import { logError } from "./errors";
+import {
+  disposeLaunchDarklyClient,
+  getLaunchDarklyClient,
+  resetFlagDefaults,
+} from "./featureFlags/client";
+import {
+  checkForExtensionDisabledReason,
+  showExtensionDisabledNotification,
+} from "./featureFlags/evaluation";
 import { constructResourceLoaderSingletons } from "./loaders";
 import { cleanupOldLogFiles, getLogFileStream, Logger, OUTPUT_CHANNEL } from "./logging";
-import { SSL_PEM_PATHS, SSL_VERIFY_SERVER_CERT_DISABLED } from "./preferences/constants";
 import { createConfigChangeListener } from "./preferences/listener";
 import { updatePreferences } from "./preferences/updates";
 import { registerProjectGenerationCommand } from "./scaffold";
 import { JSON_DIAGNOSTIC_COLLECTION } from "./schemas/diagnosticCollection";
-import { getSidecarManager } from "./sidecar";
+import { getSidecar, getSidecarManager } from "./sidecar";
 import { ConnectionStateWatcher } from "./sidecar/connections/watcher";
 import { WebsocketManager } from "./sidecar/websocketManager";
 import { getStorageManager, StorageManager } from "./storage";
@@ -101,18 +109,21 @@ const logger = new Logger("extension");
 export async function activate(
   context: vscode.ExtensionContext,
 ): Promise<vscode.ExtensionContext | undefined> {
-  observabilityContext.extensionVersion = context.extension.packageJSON.version;
+  const extVersion = context.extension.packageJSON.version;
+  observabilityContext.extensionVersion = extVersion;
   observabilityContext.extensionActivated = false;
 
-  logger.info(`Extension ${context.extension.id}" activate() triggered.`);
+  logger.info(
+    `Extension version ${context.extension.id} activate() triggered for version "${extVersion}".`,
+  );
   logUsage(UserEvent.ExtensionActivation, { status: "started" });
   try {
     context = await _activateExtension(context);
-    logger.info("Extension fully activated");
+    logger.info(`Extension version "${extVersion}" fully activated`);
     observabilityContext.extensionActivated = true;
     logUsage(UserEvent.ExtensionActivation, { status: "completed" });
   } catch (e) {
-    logger.error("Error activating extension:", e);
+    logger.error(`Error activating extension version "${extVersion}":`, e);
     // if the extension is failing to activate for whatever reason, we need to know about it to fix it
     Sentry.captureException(e);
     logUsage(UserEvent.ExtensionActivation, { status: "failed" });
@@ -147,10 +158,19 @@ async function _activateExtension(
     vscode.commands.executeCommand("confluent.showOutputChannel");
   }
 
+  // set up initial feature flags and the LD client
+  await setupFeatureFlags();
+
   // configure the StorageManager for extension access to secrets and global/workspace states, and
   // set the initial context values for the VS Code UI to inform the `when` clauses in package.json
   await Promise.all([setupStorage(), setupContextValues()]);
   logger.info("Storage and context values initialized");
+
+  // verify we can connect to the correct version of the sidecar, which may require automatically
+  // killing any (old) sidecar process and starting a new one, going through the handshake, etc.
+  logger.info("Starting/checking the sidecar...");
+  await getSidecar();
+  logger.info("Sidecar ready for use.");
 
   // set up the preferences listener to keep the sidecar in sync with the user/workspace settings
   const settingsListener: vscode.Disposable = await setupPreferences();
@@ -315,12 +335,42 @@ async function setupContextValues() {
  */
 async function setupPreferences(): Promise<vscode.Disposable> {
   // pass initial configs to the sidecar on startup
-  const configs: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration();
-  const pemPaths: string[] = configs.get(SSL_PEM_PATHS, []);
-  const trustAllCerts: boolean = configs.get(SSL_VERIFY_SERVER_CERT_DISABLED, false);
-  await updatePreferences({ tls_pem_paths: pemPaths, trust_all_certificates: trustAllCerts });
+  await updatePreferences();
   logger.info("Initial preferences passed to sidecar");
   return createConfigChangeListener();
+}
+
+/**
+ * Set up the feature flags for the extension. This includes setting the defaults, initializing the
+ * LaunchDarkly client, and checking if the extension is enabled or disabled.
+ */
+async function setupFeatureFlags(): Promise<void> {
+  // if the client initializes properly, it will set the initial flag values. otherwise, we'll use
+  // the local defaults from `setFlagDefaults()`
+  resetFlagDefaults();
+
+  const client = getLaunchDarklyClient();
+  if (client) {
+    // wait a few seconds for the LD client to initialize for the first time, because if we
+    // continue to use the client before it's ready, it will return the default values for all flags
+    const initialized = await Promise.race([
+      client
+        .waitForInitialization()
+        .then(() => true)
+        .catch((error) => {
+          logger.error("Feature flag client failed to initialize:", error);
+          return false;
+        }),
+      new Promise<boolean>((resolve) => setTimeout(() => resolve(false), 3000)),
+    ]);
+    logger.info(`Feature flag client initialization ${initialized ? "completed" : "failed"}`);
+  }
+
+  const disabledMessage: string | undefined = checkForExtensionDisabledReason();
+  if (disabledMessage) {
+    showExtensionDisabledNotification(disabledMessage);
+    throw new Error(disabledMessage);
+  }
 }
 
 /** Initialize the StorageManager singleton instance and handle any necessary migrations. */
@@ -363,12 +413,16 @@ async function setupAuthProvider(): Promise<vscode.Disposable[]> {
   // attempt to get a session to trigger the initial auth badge for signing in
   const cloudSession = await getCCloudAuthSession();
 
-  // Send an Identify event to Segment with the session info if available
+  // Send an Identify event to Segment and LaunchDarkly with the session info if available
   if (cloudSession) {
     sendTelemetryIdentifyEvent({
       eventName: UserEvent.ExtensionActivation,
       userInfo: undefined,
       session: cloudSession,
+    });
+    getLaunchDarklyClient()?.identify({
+      key: cloudSession.account.id,
+      email: cloudSession.account.label,
     });
   }
 
@@ -405,6 +459,8 @@ export function deactivate() {
   if (logStream) {
     logStream.end();
   }
+
+  disposeLaunchDarklyClient();
 
   logger.info("Extension deactivated");
 }